requirements.txt
pillow

# consider aggdraw

# dev only
black
flake8
